Update README.md
The main goal of this project is to design a wireless board using GSM technology. The main concept of this project is to design a
wireless notice board that displays various notices sent from the mobile phone. When a mobile user sends information from his
mobile, it is received by a SIM slot, which is integrated into a GSM modem at the receiver end.